refactor: added interfaces for components with child, children and onInit properties (#815)
* feat: added interfaces for components with child and children

* feat: added OnInitableComponent interface
