+++ android/beagle/src/main/java/br/com/zup/beagle/android/components/TabItem.kt
@@ -23,13 +23,15 @@ import br.com.zup.beagle.android.widget.RootView
 import br.com.zup.beagle.android.widget.WidgetView
 import br.com.zup.beagle.annotation.RegisterWidget
 import br.com.zup.beagle.core.ServerDrivenComponent
+import br.com.zup.beagle.core.SingleChildComponent
+
 @RegisterWidget
 @Deprecated(DEPRECATED_TAB_VIEW)
 data class TabItem(
     val title: String? = null,
-    val child: ServerDrivenComponent,
+    override val child: ServerDrivenComponent,
     val icon: ImagePath.Local? = null
-) : WidgetView() {
+) : WidgetView(), SingleChildComponent {
 
     @Transient
     private val viewFactory: ViewFactory = ViewFactory()
