+++ android/beagle/src/main/java/br/com/zup/beagle/android/components/layout/ScrollView.kt
@@ -25,6 +25,7 @@ import br.com.zup.beagle.core.Style
 import br.com.zup.beagle.android.widget.RootView
 import br.com.zup.beagle.android.widget.WidgetView
 import br.com.zup.beagle.annotation.RegisterWidget
+import br.com.zup.beagle.core.MultiChildComponent
 import br.com.zup.beagle.core.ServerDrivenComponent
 import br.com.zup.beagle.widget.core.Flex
 import br.com.zup.beagle.widget.core.FlexDirection
@@ -32,11 +33,11 @@ import br.com.zup.beagle.widget.core.ScrollAxis
 
 @RegisterWidget
 data class ScrollView(
-    val children: List<ServerDrivenComponent>,
+    override val children: List<ServerDrivenComponent>,
     val scrollDirection: ScrollAxis? = null,
     val scrollBarEnabled: Boolean? = null,
     override val context: ContextData? = null
-) : WidgetView(), ContextComponent {
+) : WidgetView(), ContextComponent, MultiChildComponent {
 
     @Transient
     private val viewFactory: ViewFactory = ViewFactory()
