+++ schema/kotlin-core/src/main/kotlin/br/com/zup/beagle/core/AccessibilityComponent.kt
@@ -18,9 +18,8 @@ package br.com.zup.beagle.core
 
 /**
  * Component that hold the accessibility
- * @property accessibility
- *                              attribute will enable a textual information to explain the view content
- *                              in case a screen reader is use
+ * @property accessibility attribute will enable a textual information to explain
+ * the view content in case a screen reader is use
  */
 interface AccessibilityComponent: ServerDrivenComponent {
     val accessibility: Accessibility?
