+++ schema/kotlin-core/src/main/kotlin/br/com/zup/beagle/core/StyleComponent.kt
@@ -17,7 +17,7 @@
 package br.com.zup.beagle.core
 
 /**
- * Component that hold the appearance
+ * Component that hold the style
  * @property style attribute will enable a few visual options to be changed.
  */
 interface StyleComponent : ServerDrivenComponent {
